:sparkles: Check inventory on change world
