+++ src/main/java/com/github/euonmyoji/epicbanitem/check/listener/InventoryListener.java
@@ -9,14 +9,13 @@ import com.github.euonmyoji.epicbanitem.check.CheckRuleService;
 import com.github.euonmyoji.epicbanitem.check.Triggers;
 import com.github.euonmyoji.epicbanitem.util.NbtTagDataUtil;
 import com.github.euonmyoji.epicbanitem.util.TextUtil;
+import com.google.common.collect.Streams;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
-
-import com.google.common.collect.Streams;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.block.BlockSnapshot;
 import org.spongepowered.api.data.DataContainer;
@@ -216,39 +215,6 @@ public class InventoryListener {
                 event.setCancelled(true);
             }
         }
-    }
-
-    @Listener
-    @SuppressWarnings({"OptionalGetWithoutIsPresent", "UnstableApiUsage"})
-    public void onJoin(Join event, @First Player player) {
-        Streams.stream(service
-            .checkInventory(player.getInventory(), player.getWorld(), Triggers.JOIN, player))
-            .filter(tuple -> tuple.getFirst().isBanned())
-            .forEach(
-                tuple -> {
-                    tuple
-                        .getFirst()
-                        .getFinalView()
-                        .map(dataContainer -> NbtTagDataUtil.toItemStack(dataContainer, tuple.getSecond().peek().get().getQuantity()))
-                        .ifPresent(
-                            finalItem -> {
-                                CheckResult checkResult = tuple.getFirst();
-                                Inventory inventory = tuple.getSecond();
-                                ItemStack itemStack = inventory.peek().get();
-                                inventory.set(finalItem);
-                                TextUtil
-                                    .prepareMessage(
-                                        Triggers.JOIN,
-                                        TextUtil.getDisplayName(itemStack),
-                                        TextUtil.getDisplayName(finalItem),
-                                        ((CheckResult.Banned) checkResult).getBanRules(),
-                                        checkResult.isUpdateNeeded()
-                                    )
-                                    .forEach(player::sendMessage);
-                            }
-                        );
-                }
-            );
     }
 
     @Listener(order = Order.FIRST, beforeModifications = true)
